! feat: add spin circle
