+++ scripts/webpack.base.config.js
@@ -37,7 +37,7 @@ const createHappypackPlugin = () => {
         loader: 'sass-loader',
         options: {
           // data: "$icon-base-url: '//yuanzhaohao.github.io/dashkit-fonts';",
-          data: "$icon-base-url: '//localhost:8999/static/fonts';",
+          data: "$icon-base-url: '/static/fonts';",
         },
       },
     ]),
